Moved createAnalyzeFunction to a separate file.
