+++ packages/seo-integration/src/analysis.js
@@ -1,7 +1,9 @@
 import { applyFilters } from "@wordpress/hooks";
 import { __ } from "@wordpress/i18n";
 import { isObject } from "lodash";
-import { AnalysisWorkerWrapper, createWorker, Paper } from "yoastseo";
+import { AnalysisWorkerWrapper, createWorker } from "yoastseo";
+
+import createAnalyzeFunction from "./analyze";
 
 /**
  * Creates the analysis worker.
@@ -33,112 +35,6 @@ const createAnalysisConfiguration = ( configuration = {} ) => {
 };
 
 /**
- * Creates a paper based on the given data, keyphrase and configuration.
- *
- * @param {Object} data The paper data and metadata, e.g. content, title, date, slug.
- * @param {Object} keyphrase The keyphrase to use for the analysis.
- * @param {Object} configuration Extra configuration.
- *
- * @returns {Paper} A paper that can be analyzed using the web worker.
- */
-function createPaper( data, keyphrase, configuration ) {
-	return new Paper(
-		data.content,
-		{
-			// Keyphrase data.
-			keyword: keyphrase.keyphrase,
-			synonyms: keyphrase.synonyms,
-			// General data and metadata.
-			description: data.metadescription,
-			title: data.seoTitle,
-			titleWidth: data.seoTitleWidth,
-			permalink: data.slug,
-			date: data.date,
-			// Configuration data.
-			locale: configuration.locale,
-		},
-	);
-}
-
-/**
- * Adapts the object with related keyphrases to a a structure
- * that the analysis web worker is able to consume.
- *
- * @param {Object} relatedKeyphrases The related keyphrases.
- *
- * @returns {Object} The transformed related keyphrases.
- */
-function adaptRelatedKeyphrases( relatedKeyphrases ) {
-	const transformed = {};
-
-	Object.keys( relatedKeyphrases ).forEach(
-		id => {
-			transformed[ id ] = {
-				keyword: relatedKeyphrases[ id ].keyphrase,
-				synonyms: relatedKeyphrases[ id ].synonyms,
-			};
-		},
-	);
-
-	return  transformed;
-}
-
-/**
- * Adapt the results from the analysis to the structure
- * that the SEO store expects.
- *
- * @param {Object} results The results returned by the analysis web worker.
- *
- * @returns {Object} The adapted results.
- */
-function adaptResults( results ) {
-	const { seo, readability } = results.result;
-
-	// Split out the SEO result of the main focus keyphrase.
-	const focusKeyphraseResults = seo[ "" ];
-	delete seo[ "" ];
-
-	return {
-		seo: {
-			focus: focusKeyphraseResults,
-			...seo,
-		},
-		readability,
-	};
-}
-
-/**
- * Creates a callback function to trigger a new analysis
- * based on the given analysis web worker and configuration.
- *
- * @param {AnalysisWorkerWrapper} worker The web worker wrapper.
- * @param {Object} configuration Configuration.
- *
- * @returns {function} The analysis callback function.
- */
-function createAnalyzeFunction( worker, configuration ) {
-	/**
-	 * A callback function that analyzes the data from the SEO store.
-	 *
-	 * @param {Object} data The data from the SEO store.
-	 * @param {Object} keyphrases The keyphrases to analyze.
-	 * @param {Object} [config] Optional extra configuration to use.
-	 *
-	 * @returns {Object} The results of the analysis.
-	 */
-	return async ( data, keyphrases, config = {} ) => {
-		const { focus, ...relatedKeyphrases } = keyphrases;
-
-		const results = await worker.analyzeRelatedKeywords(
-			createPaper( data, focus, configuration ),
-			adaptRelatedKeyphrases( relatedKeyphrases ),
-		);
-
-		return adaptResults( results );
-	};
-}
-
-/**
  * Creates the analysis worker.
  *
  * @param {string} workerUrl The URL of the analysis worker.
@@ -147,7 +43,7 @@ function createAnalyzeFunction( worker, configuration ) {
  *
  * @returns {function} The analysis worker wrapper.
  */
-const createAnalysisWorker = async( { workerUrl, dependencies, configuration = {} } ) => {
+const createAnalysisWorker = async ( { workerUrl, dependencies, configuration = {} } ) => {
 	const worker = createAnalysisWorkerWrapper( { workerUrl, dependencies } );
 
 	try {
